Set gRedCoinsCollected
network_receive_collect_coin will increment the gRedCoinsCollected variable if the coin was collected in a different area (In Shifting Sand Land while in the Pyramid, for instance)

Also plays a red coin sound to be consistent with the yellow coin sounds playing from out of area. Though I'm not sure if this is the most appropriate place for that.